sandbox/plaxchat2/requirements.txt
faiss-cpu
openai
sqlite-utils
pydantic
rapidfuzz
python-dotenv
